* misc: lots of changes to make YAM compilable by SAS/C again. Mainly some   typecasts were added. The biggest change happened in the toolbar subclasses,   because SAS/C is not able to handle the tr() calls within a structure   definition.

+++ src/classes/InfoBar.c
@@ -199,7 +199,7 @@ DECLARE(SetFolder) // struct Folder *newFolder
 {
   GETDATA;
   struct Folder *folder = msg->newFolder;
-  LONG result = -1;
+  ULONG result = (ULONG)-1;
 
   ENTER();
